+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 
 - Fix variable name in quickstart example [#316](https://github.com/stac-utils/pystac-client/pull/316)
+- `StacApiIO.write_text_to_href` [#312](https://github.com/stac-utils/pystac-client/pull/312)
 
 ## Removed